+++ panoptes_cli/scripts/panoptes.py
@@ -36,3 +36,4 @@ from panoptes_cli.commands.configure import *
 from panoptes_cli.commands.project import *
 from panoptes_cli.commands.subject import *
 from panoptes_cli.commands.subject_set import *
+from panoptes_cli.commands.workflow import *